Langchain /Day 04/requirements.txt
langchain 
python-dotenv 
streamlit 
langchain_google_genai 
langchain-ollama
langchain_community
